Modified Server to add dummy data each minute. Implemented on client to call server for current humidity

+++ airmes/src/main/kotlin/com/maja/airmes/dtos/AirSensorDto.kt
@@ -12,5 +12,5 @@ class AirSensorDto(
         val id: Long,
         val time: Timestamp,
         val humidity: Int?,
-        val temperature: Int?
+        val temperature: Double?
 )
